# Optimize block decoding