Use memory copies when nulls aren't present