@tdcmeehan 2020-02-10
Added a quick benchmark for `BIGINT`.

Before:

```
Benchmark                                      (channelCount)  (hasNull)  (type)  Mode  Cnt   Score   Error  Units
BenchmarkLongArrayBlockEncoding.testReadBlock               1       true  BIGINT  avgt   30  47.135 ± 1.465  us/op
BenchmarkLongArrayBlockEncoding.testReadBlock               1      false  BIGINT  avgt   30  38.394 ± 0.834  us/op
BenchmarkLongArrayBlockEncoding.testReadBlock               2       true  BIGINT  avgt   30  72.875 ± 1.708  us/op
BenchmarkLongArrayBlockEncoding.testReadBlock               2      false  BIGINT  avgt   30  57.430 ± 1.962  us/op
```

After:

```
Benchmark                                      (channelCount)  (hasNull)  (type)  Mode  Cnt   Score   Error  Units
BenchmarkLongArrayBlockEncoding.testReadBlock               1       true  BIGINT  avgt   30  30.400 ± 1.440  us/op
BenchmarkLongArrayBlockEncoding.testReadBlock               1      false  BIGINT  avgt   30  12.430 ± 0.742  us/op
BenchmarkLongArrayBlockEncoding.testReadBlock               2       true  BIGINT  avgt   30  50.845 ± 2.144  us/op
BenchmarkLongArrayBlockEncoding.testReadBlock               2      false  BIGINT  avgt   30  19.241 ± 1.037  us/op
```

So on the microbenchmark we see a 30% improvement with nulls, 66% improvement without nulls.