@tdcmeehan 2020-04-10
Once #14329 this should improve performance on the raw page format further, will benchmark.