@tdcmeehan 2020-06-16
Performance results for HiveFileFormatBenchmark on `PRESTO_PAGE`:  https://gist.github.com/tdcmeehan/a102aaf5cb8be789d80c62b6df17e6c5

Tl;dr, most dramatic improvement is seen for primitives, maps of primitives also see speedup, no significant movement on variable width types (as this is not optimized).