@tdcmeehan 2020-06-17
Optimizations for writes can go on a different PR.