@tdcmeehan 2020-06-17
@yingsu00 how about we add `readInts`\`readLongs`\etc method to `SliceInput`?  That way we could avoid the intermediate slice creation and directly use unsafe to perform a memory copy.