@yingsu00 2020-06-21
> @yingsu00 how about we add `readInts` \ `readLongs` \ etc method to `SliceInput`? That way we could avoid the intermediate slice creation and directly use unsafe to perform a memory copy.

That'll be nice. But we need to fork Slice first :)